Added debouncer and fixed bugs
I added a debouncer with a 30ms delay to save compute when the user's quickly entering text: it won't recompute it on every keystroke, but on the input 30ms after the last keystroke.

- I'm not sure if this is long enough to make an impactful difference (time between keystrokes?)

I also fixed the bug that occurred when I entered a word like "twelve" that's considered a "Value" by compromise.js; previously, it would directly return the word without placing it inside two arrays. This meant that it'd be taking the first letter of the word {and would try to get the first index *within* the first letter (which doesn't exist)} and using that.