Updated Indexer
Update on the rewrite:
The tokenizer works, the index works, it's a lot more legible than it was before, I burnt out a lot of the bloat, and I've added comments.

I'm yet to implement proper latex tokenization, proper codeblock handling and wordcounts in a document, but I'm going to work on those soon.