Went on a rampage and deleted the lemmatizer
I'm not sure if this was smart, but I can always add it back in later (considering that I have a functioning implementation in the original search-engine branch). I'll probably end up doing just that before the refactor's out, but I don't want to deal with the hassle of implementing it on the backend right now.

I also connected the index to the frontend; it'll show the documents every word comes from now.