Update refactored_search.js
I have a search system working; this is awesome.
I don't have any insight into why it's giving me the results it's giving (beyond knowing that it's summing idf scores); not so awesome.

I need to figure out why that's the case, and I need to include the document titles in the searchable systems. I also need to separate the tags.