Replaced <input> element with <div>
Input elements don't permit innerHTML; I replaced the input with a content editable div to allow me to place <span> tags within. This will allow me to bubble out tags when they're entered by the user; I'll also gain greater control over the styling of the text.

I need to style the text of the div (the color & font, mainly) to match that of the original input element. I also need to figure out how to have placeholder text that disappears when edited.